+++ client/src/components/layout/MobileSidebar.js
@@ -1,0 +1,49 @@
+import PropTypes from 'prop-types';
+import {
+  Dialog,
+  DialogBackdrop,
+  DialogPanel,
+  TransitionChild,
+} from '@headlessui/react';
+import { XMarkIcon } from '@heroicons/react/24/outline';
+import Sidebar from './Sidebar';
+
+function MobileSidebar({ open, onClose }) {
+  return (
+    <Dialog open={open} onClose={onClose} className="relative z-50 lg:hidden">
+      {/* Backdrop */}
+      <DialogBackdrop
+        transition
+        className="fixed inset-0 bg-gray-900/80 transition-opacity duration-300 ease-linear data-[closed]:opacity-0"
+      />
+
+      <div className="fixed inset-0 flex">
+        {/* Sliding panel */}
+        <DialogPanel
+          transition
+          className="relative mr-16 flex w-full max-w-xs flex-1 transform transition duration-300 ease-in-out data-[closed]:-translate-x-full"
+        >
+          {/* Close button */}
+          <TransitionChild>
+            <div className="absolute left-full top-0 flex w-16 justify-center pt-5 duration-300 ease-in-out data-[closed]:opacity-0">
+              <button type="button" onClick={onClose} className="-m-2.5 p-2.5">
+                <span className="sr-only">Close sidebar</span>
+                <XMarkIcon className="h-6 w-6 text-white" aria-hidden="true" />
+              </button>
+            </div>
+          </TransitionChild>
+
+          {/* Reuse Sidebar content */}
+          <Sidebar />
+        </DialogPanel>
+      </div>
+    </Dialog>
+  );
+}
+
+MobileSidebar.propTypes = {
+  open: PropTypes.bool.isRequired,
+  onClose: PropTypes.func.isRequired,
+};
+
+export default MobileSidebar;